refactor: style component file name #29
