+++ src/components/footer/styles/Footer.styled.ts
@@ -1,4 +1,4 @@
 import styled from "styled-components";
 
-export const FooterContainer= styled.div`
+export const FooterContainer = styled.div`
 `;
