+++ src/components/short-clip-video/styles/ShortClipVideo.styled.ts
