+++ src/components/trending-video/styles/TrendingVideo.styled.ts
