+++ src/components/video-item/styles/VideoItem.styled.ts
