+++ src/layouts/auth-layout/styles/AuthLayout.styled.ts
