+++ src/layouts/main-layout/styles/MainLayout.styled.ts
