+++ src/utils/components/divider/styles/Divider.styled.ts
