+++ src/utils/components/popper/index.ts
@@ -1,5 +1,5 @@
 import Popper from './Popper';
-import PopperContent from './components/popper-item-content/PopperItemContent';
+import PopperContent from './popper-item-content/PopperItemContent';
 
 export {
   Popper,
