+++ src/views/nft-mode/styles/NftMode.styled.ts
