+++ src/views/playlist/components/playlist-item/components/playlist-video/PlaylistVideo.tsx
@@ -8,7 +8,7 @@ import {
   PlaylistVideoTitleContainer,
   PlaylistVideoTitleLabel,
   PlaylistVideoVideosContainer
-} from './styles/PlaylistVideo.styled.component';
+} from './styles/PlaylistVideo.styled';
 
 interface PlaylistVideoProps {
   videos: VideoModel[];
