+++ src/views/playlist/components/playlist-item/components/playlist-video/styles/PlaylistVideo.styled.ts
