+++ src/views/playlist/components/playlist-item/components/publication-dropdown-item/PublicationDropdownContent.tsx
@@ -7,7 +7,7 @@ import {
   PublicationDropdownItemNameContainer,
   PublicationDropdownItemNameLabel,
   PublicationDropdownItemInfoLabel
-} from './styles/PublicationDropdownContent.styled.component';
+} from './styles/PublicationDropdownContent.styled';
 
 interface PublicationDropdownItemProps {
   label: string;
