+++ src/views/upload-video/components/additional-content/styles/AdditionalContent.styled.ts
