+++ src/views/upload-video/styles/UploadVideo.styled.ts
