+++ src/views/video-detail/components/video-recommendation/styles/VideoRecommendation.styled.ts
