+++ src/views/video-detail/styles/VideoDetail.styled.ts
